requirements.txt
transformers
torch
sentence_transformers
python-dotenv
altair<5
streamlit
